+++ include/librealsense2/rs.h
@@ -22,6 +22,7 @@ extern "C" {
 #include "h/rs_processing.h"
 #include "h/rs_record_playback.h"
 #include "h/rs_sensor.h"
+#include "h/rs_aus.h"
 
 #define RS2_API_MAJOR_VERSION    2
 #define RS2_API_MINOR_VERSION    50
@@ -52,14 +53,6 @@ extern "C" {
 * \return size of rs2_raw_data_buffer
 */
 int rs2_get_raw_data_size(const rs2_raw_data_buffer* buffer, rs2_error** error);
-
-
-//AUS SAMER
-int rs2_get_aus_counters_names_size(const rs2_aus_counters_names* buffer, rs2_error** error);
-void rs2_delete_aus_counters_names(const rs2_aus_counters_names* buffer);
-const char ** rs2_get_aus_counters_names_data(const rs2_aus_counters_names* buffer, rs2_error** error);
-const char * rs2_get_aus_counter_name_data(const rs2_aus_counters_names* buffer, int i, rs2_error** error);
-
 
 /**
 * Delete rs2_raw_data_buffer
@@ -102,15 +95,7 @@ void rs2_reset_logger( rs2_error ** error);
 */
 void rs2_enable_rolling_log_file( unsigned max_size, rs2_error ** error );
 
-//AUS
-void rs2_aus_set(const char* counter, int value, rs2_error** error);
-void rs2_aus_increase(const char * counter, rs2_error** error);
-int rs2_aus_get_counter(const char* counter, rs2_error** error);
-void rs2_aus_start_timer(const char* timer, rs2_error** error);
-void rs2_aus_stop_timer(const char* timer, rs2_error** error);
-rs2_time_t rs2_aus_get_timer(const char* timer, rs2_error** error);
-const rs2_aus_counters_names* rs2_aus_get_counters_names(rs2_error** error);
-//END OF AUS
+
 
 unsigned rs2_get_log_message_line_number( rs2_log_message const * msg, rs2_error** error );
 const char * rs2_get_log_message_filename( rs2_log_message const * msg, rs2_error** error );
