+++ include/librealsense2/rs.hpp
@@ -13,6 +13,7 @@
 #include "hpp/rs_record_playback.hpp"
 #include "hpp/rs_sensor.hpp"
 #include "hpp/rs_pipeline.hpp"
+#include "hpp/rs_aus.hpp"
 
 namespace rs2
 {
@@ -51,92 +52,7 @@ namespace rs2
         error::handle( e );
     }
     
-    //AUS SAMER
-
-    inline void aus_set(const char* counter, int value = 0)
-    {
-        rs2_error* e = nullptr;
-        rs2_aus_set(counter, value, &e);
-        error::handle(e);
-    }
-
-    inline void aus_increase(const char* counter)
-    {
-        rs2_error* e = nullptr;
-        rs2_aus_increase(counter, &e);
-        error::handle(e);
-    }
-
-    inline int aus_get_counter(const char* counter)
-    {
-        rs2_error* e = nullptr;
-        int result = rs2_aus_get_counter(counter, &e);
-        error::handle(e);
-        return result;
-    }
-
-    inline void aus_start_timer(const char* timer)
-    {
-        rs2_error* e = nullptr;
-        rs2_aus_start_timer(timer, &e);
-        error::handle(e);
-    }
-
-    inline void aus_stop_timer(const char* timer)
-    {
-        rs2_error* e = nullptr;
-        rs2_aus_stop_timer(timer, &e);
-        error::handle(e);
-    }
-
-    inline rs2_time_t aus_get_timer(const char* timer)
-    {
-        rs2_error* e = nullptr;
-        rs2_time_t result = rs2_aus_get_timer(timer, &e);
-        error::handle(e);
-        return result;
-    }
-
-    inline std::vector<std::string> aus_get_counters_names()
-    {
-        
-        rs2_error* e = nullptr;
-        auto buffer = rs2_aus_get_counters_names(&e);
-        std::shared_ptr<const rs2_aus_counters_names> list(buffer, rs2_delete_aus_counters_names);
-        error::handle(e);
-
-        auto size = rs2_get_aus_counters_names_size(list.get(), &e);
-        error::handle(e);
-
-        std::vector<std::string> results;
-        for (size_t i = 0; i < size; ++i) {
-            const char * item_cstr = rs2_get_aus_counter_name_data(list.get(), i, &e);
-            error::handle(e);
-            results.push_back(item_cstr);
-        }
-        //auto start = rs2_get_aus_counters_names_data(list.get(), &e);
-
-        return results;
-    }
-
-    /*
-            std::vector<uint8_t> results;
-
-            rs2_error* e = nullptr;
-            auto buffer = rs2_build_debug_protocol_command(_dev.get(), opcode, param1, param2, param3, param4,
-                (void*)data.data(), (uint32_t)data.size(), &e);
-            std::shared_ptr<const rs2_raw_data_buffer> list(buffer, rs2_delete_raw_data);
-            error::handle(e);
-
-            auto size = rs2_get_raw_data_size(list.get(), &e);
-            error::handle(e);
-
-            auto start = rs2_get_raw_data(list.get(), &e);
-            error::handle(e);
-
-            results.insert(results.begin(), start, start + size);
     
-    */
 
     /*
         Interface to the log message data we expose.
